fix(build): pre-existing macOS userspace build breaks before P1 starts
P1's host test (scripts/test.sh host) cannot pass on macOS without these
fixes. Discovered while wiring the host subcommand (T3); split out so the
T3 commit stays narrowly scoped.

These are pure userspace-build hygiene; no behavior change to the kernel
build path. P2 (KP-port audit) will revisit deeper concerns; this commit
just unblocks the macOS-side test harness.

- include/arch/arm64/pgtable.h: wrap the entire kernel PTE/TLB section
  in #ifndef __USERSPACE__. linux/types.h and friends are unavailable on
  macOS; inline.c gates all write_insts_* call sites with the same macro,
  so the userspace build needs no PTE infrastructure.

- src/arch/arm64/inline.c: guard #include <linux/set_memory.h> behind
  #ifndef __USERSPACE__. macOS has no Linux kernel headers.

- src/platform/hmem_user.c: change rox_set_memory* / rw_set_memory_nop
  signatures from uint64_t addr to uintptr_t addr to match the
  hook_mem_ops_t typedef. On macOS arm64, uint64_t = unsigned long long
  and uintptr_t = unsigned long are distinct types under -Werror.

- src/platform/darwin.c: split the darwin_alloc helper into top-level
  platform_alloc_rox / platform_alloc_rw definitions for clarity, and
  expand the comment on platform_set_rw to document why VM_PROT_COPY is
  required (mprotect cannot re-add PROT_WRITE after page is RX without
  max_prot allowing it; VM_PROT_COPY produces a CoW writable page).
  No functional change to vm_protect call.

- CMakeLists.txt: add -mbranch-protection=standard to the kernelhook
  target. The transit asm stub in src/arch/arm64/transit.c memcpy's the
  compiled _transit / _fp_transit prologues into the transit buffer at
  transit[2]. The asm uses 'sub x16, x16, #12' relative to an 'adr x16, .'
  at transit[3] to locate the self-pointer slot at transit[0..1]. That
  arithmetic only holds if the compiler emits a BTI C landing pad as the
  first instruction of _transit (occupying transit[2] in the buffer copy).
  Without -mbranch-protection=standard, no BTI C is emitted, the adr
  lands one slot earlier, and the self-pointer load is corrupt → crash
  on first hook invocation. CheckCCompilerFlag-gated to avoid breaking
  toolchains that don't support the flag.

+++ src/platform/hmem_user.c
@@ -58,14 +58,14 @@ static void rox_free(void *ptr)
         platform_free(ptr, size);
 }
 
-static int rox_set_memory(uint64_t addr, int numpages, int (*fn)(uint64_t, uint64_t))
+static int rox_set_memory(uintptr_t addr, int numpages, int (*fn)(uint64_t, uint64_t))
 {
-    return fn(addr, (uint64_t)numpages * platform_page_size());
+    return fn((uint64_t)addr, (uint64_t)numpages * platform_page_size());
 }
 
-static int rox_set_memory_rw(uint64_t addr, int numpages) { return rox_set_memory(addr, numpages, platform_set_rw); }
-static int rox_set_memory_ro(uint64_t addr, int numpages) { return rox_set_memory(addr, numpages, platform_set_ro); }
-static int rox_set_memory_x(uint64_t addr, int numpages)  { return rox_set_memory(addr, numpages, platform_set_rx); }
+static int rox_set_memory_rw(uintptr_t addr, int numpages) { return rox_set_memory(addr, numpages, platform_set_rw); }
+static int rox_set_memory_ro(uintptr_t addr, int numpages) { return rox_set_memory(addr, numpages, platform_set_ro); }
+static int rox_set_memory_x(uintptr_t addr, int numpages)  { return rox_set_memory(addr, numpages, platform_set_rx); }
 
 /* ---- RW ops wrappers ---- */
 
@@ -84,7 +84,7 @@ static void rw_free(void *ptr)
         platform_free(ptr, size);
 }
 
-static int rw_set_memory_nop(uint64_t addr __maybe_unused, int numpages __maybe_unused)
+static int rw_set_memory_nop(uintptr_t addr __maybe_unused, int numpages __maybe_unused)
 {
     return 0;
 }
